src/bundle: do not attempt to take ownership if running as non-root
As part of the plain bundle ownership safeguard, RAUC attempts to chown
bundles that are neither owned by root nor by the uid executing rauc to
user root.

This cannot work when rauc runs as non-root. Thus emit a more helpful
error message and abort in advance.

To not call geteuid() unnecessary often, introduce variable euid to
store the result.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>