add meson support
This adds support to build RAUC source code and documentation using
meson/ninja.

The compiler and linker flags were adapted from automake to match the
current behavior as good as possible.

The minimum required meson version is 0.51 as a compromise between
availability of useful features and support for host systems not having
latest and greatest meson available (0.64 at the time of writing).

Note that some used features are deprecated already in latests meson,
but used since their replacements were not available in 0.51, yet.

Feature used of 0.51:

* get_variable() method of dependency objects (replaces
  get_pkgconfig_variable() that is deprecated with meson 0.56)

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>