test/stats: make tests involving floating point calculations more lax
On i387 floating point calculations can be quite surprising. The
motivation for this change came from the observation that compiling and
testing rauc-1.8 on the i386 Debian build machine had a failure on

	g_assert_cmpfloat(r_stats_get_avg(stats), ==, 62.54615384615385);

. In that case r_stats_get_avg(stats) (when stored in a double typed
variable) returned

	0x1.f45e85e85e85fp+5

which is 0x1f45e85e85e85f00 / 2**55. However when stored in a long
double typed variable it has the value

	0x1.f45e85e85e85e85p+5

This value is 0x1f45e85e85e85e85 / 2**55 and so it has a significand
that cannot be stored in the 53 bits that are allocated for double typed
variables. 62.54615384615385 is a plain double and so evaluates to
0x1.f45e85e85e85fp+5.

This makes the comparison implemented in g_assert_cmpfloat() (which uses
long double precision) fail. To fix that, use
g_assert_cmpfloat_with_epsilon() and also adapt the other tests that
involve non-trivial calculations to this approach.

As of now rauc depends on glib >= 2.50, so add a conditional backport of
the two required macros that are available only in glib 2.58.

Fixes: https://github.com/rauc/rauc/issues/996
Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>