src/nbd: be lenient with libcurl without proxy support
In streaming mode, RAUC currently assumes that libcurl is built with proxy
suport. When configuring libcurl, it tries to set CURLOPT_HTTPPROXYTUNNEL which
will fail when using libcurl built without proxy support with
CURLE_UNKNOWN_OPTION (48). In that case, we don't need to set
CURLOPT_HTTPPROXYTUNNEL, as no proxy will be used anyway.

CURLOPT_SUPPRESS_CONNECT_HEADERS is always available, so we don't need to
handle it the same way.

Signed-off-by: Christian Meusel <christian.meusel@posteo.de>
Signed-off-by: Jan Luebbe <jlu@pengutronix.de>