manifest: handle convert/converted image properties
In addition to single-file artifacts, we want to support directory
trees, as they are useful for containers (among many other things).
As these files' meta-data (owner, group, permissions, capabilities, ...)
is important, we need to preserve it during bundle creation.

Instead of using directory trees in the input bundle directory and
requiring the user to setup a (fake) root environment, we keep using tar
files as input. They can be generated easily from many build systems and
include all the necessary meta-data.

The convert= property can then be used to select one or more conversion
methods to use during 'rauc bundle'. The resulting file/directory name
is then stored in the converted= property in the same order. In a
bundle, the properties must have the same length and order.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>