data: allow skipping service files installation
When 'service' option is disabled, there is no point in installing
service files.
Since currently all files installed from data/ are service-related,
simply use subdir_done() here.

This is something that is wrong in autotools, too.
But since we intend to remove autotools support in the near future, fix
this for meson only.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>