src/main: Split help into bundle handling and target commands
The list of RAUC commands increased over the last releases.
Together with the dual use of rauc as host and target tool it becomes
harder to distinct between host and target tool usage.

While for some commands (as 'rauc info') no clear category can be found,
for others (as 'rauc service' or 'rauc status') the distinction is quite
clear.

To better guide the user, we split the subcommand info prompt into
'bundle handling' and 'target' subcommands.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>