src/config_file: simplify return handling of load_config()
The `free` exit label in load_config() did nothing more than setting
the return argument *config to either NULL on failure or to the stolen
pointer of the config struct on success.

We can safely set *config = NULL unconditionally in the beginning and
the override this at the end in case of success.

Having this, there is no need for the 'free' label anymore and we can
safely replace all:

  res = FALSE;
  goto free;

statements by an emmediate return:

  return FALSE;

Then we can safely remove remaining uses of the `res` variable.

This shortens and simplifies the flow of this large function and is
probably also less error-prone.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>