src: add installation transaction UUID
It can either be provided manually (via D-Bus or command line arg) or
will be auto-generated.

The transaction UUID allows us to identify and track an installation
run.

It is an important building block for adding future features like
history logging, life cycle tracing, and more.

First of all, add a simple printout of the UUID when the installation
starts so we can follow it in the logs.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>