src/context: consistently return allocated string in get_cmdline_bootname()
So far, except for "_external_", all return values of
get_cmdline_bootname() were allocated string. But the function state to
return a const gchar and in r_context_configure_target() we g_strdup()
the return value directly.

Be consistent and avoid leakage by always returning an allocated string
and using this directly.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>