src: split determine_slot_states() and update_external_mount_points()
The slot state determined in determine_slot_states() normally never
changes on a system during the runtime of the RAUC service.

So far, RAUC has called determine_slot_states() again on each install
and status update call. This is not only unnecessarily often, it is
later than necessary because the information required to determine the
slot states is available quite early (after having read config file and
cmdline).

The only things that may change during the runtime of RAUC, are the
external mount points. But, since we have already moved this part to a
distinct method and made this public, we can now remove this from
determine_slot_states() and call it explicitly where needed.

The determine_slot_states() method we can then safely move into early
service startup.
Former calls of determine_slot_states() in do_install_bundle() and
create_slotstatus_array() can now be replaced by calls to
update_external_mount_points().
The call in r_on_handle_mark() can entirely be removed since mount
information is not relevant there.

An explicit call to determine_slot_states() needs to be re-added for
install_start() in the non-service case.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>