install: refactor get_install_images to produce image installation plans
During installation, we keep looking up the same information (such as the
target RaucSlot) multiple times. By using additional structs with the life-time
of one installation process, we can keep this information once it is known.

This commit only changes from a GList of RaucImage pointers to a GPtrArray of
RImageInstallPlan and leaves the rest as is.

Follow-up commits will use the RImageInstallPlan structure to store information
relevant to the installation of one image. This will also make it easier to
handle images installation into artifact repos. Finally, it will allow better
progress output, as we can estimate the amount of work to perform earlier.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>