src: give check_slot low progress weight
This adds a function for formatted weighting of progress and uses it to
give the 'check_slot' progress sub step of 'update_slots' a weight of 1
while the actual 'copy_image' step gets a weight of 9 to reflect that in
'check_slot' hardly any work is done in contrast to 'copy_image'.

An alternative could be to just fully remove the check_slot sub step
which does not check nearly as much as the name might indicate.

Without this patch and with having a single slot, progress will be:

| 40% Updating slots
| 40% Checking slot rootfs.1
| 70% Checking slot rootfs.1 done.
| 70% Copying image to rootfs.1
| 99% Copying image to rootfs.1 done.

Note that this almost instantly jumps from 40% to 70%.

With the patch instead, we get:

| 40% Updating slots
| 40% Checking slot rootfs.1
| 46% Checking slot rootfs.1 done.
| 46% Copying image to rootfs.1
| 99% Copying image to rootfs.1 done.
| 99% Updating slots done.

Which better reflects the time actually spent in each subtask.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>