src/install: fix string memory leaks
install_args_update() was sometimes called with a static string and sometimes
with a string allocated using g_strdup_printf(). In the latter case, this
string was not freed.

Instead move the formatting into install_args_update(). As it needs its own
copy anyway, this avoids the confusing alternative of a consuming the
reference.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>