src: use system status to determine boot or service restart
With the boot-id information from the system status file, the service
can determine if the system was (re)booted.

This works by comparing the current boot ID with the ID stored in the
status file. If they differ, we know a reboot occurred and then store the
current boot ID in the status file.
A subsequent restart of the service would the see the current and the
stored boot ID being equal and just report a service restart.

We emit a normal log message for now but the plan is to use this later
for dedicated logging of reboots.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>