src/context: stop useless mixing of input env and output variables
launch_and_wait_variables_handler() takes a GHashTable with environment
variables as input. The same hash table is also used to emit variables
obtained from the handler.

This is a bit confusing and useless, since there is no piece of code
that passes any input to the handler.

Rework the function to explicitly return a newly allocated hash table on
success.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>