install: allow to skip bundle verification
During development, bundles are often signed by self signed certificates
which change e.g. with every run of the CI.

To test such bundles on existing systems, it is required to add the
used certificate to /etc/rauc/ca.cert.pem.  This is cumbersome because
/etc might be read-only or the used certificate might not be available.

Patch adds a '--no-verify' option to 'install' which allows to skip
bundle verification.

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>