install: disable '--no-verify' support by default
Allowing to disable bundle verification can pose a security risk on
badly configured systems.  E.g. when applications lack proper input
validation and pass filenames directly to 'rauc install'; or when
D-Bus messages are not filtered.

To reduce impact of such problems, it is required that the support for
'--no-verify' is explicitly activated during configure time.

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>