src/main: add missing --key argument for info, extract, extract-signature
For the decryption use case, this specifies the decryption key.
The global --key options has been replaced by command-specific ones but
the purpose of '--key' for 'rauc info' and the others has not been
considered.

Since the former global options are still available but hidden, using
the options always worked but it has not been shown in help output or
man page.

Fixes: 26026d6b ("src/main: expose --cert and --key arguments for
signing commands only").

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>