src/install: fix asymmetric "update_slots" step termination
The "update_slots" step is started *after* marking the boot slot bad,
right *before* actually handling the installation plans.

For a symmetric termination of this step, it should thus be terminated
right *after* handling the installation plans and *before* marking the
slots good again.

However, it is called *after* marking the slots good again.

This is inconsistent and also complicates label handling. Fix both by
making it symmetric.

While at it, do the same for the install_args_update() calls.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>