src/mark: do not let mark_active() fail on failed status update
Currently, the case that writing the status file fails is handled
inconsistently. In mark_run() we ignore the returned error
(R_INSTALL_ERROR_FAILED) while in launch_and_wait_default_handler() we
make it fatal and let the installation fail.

The latter case is problematic since it does not match the reality:
Even if an error is reported, the actual installation was successful and
a reboot will lead us into the newly installed system.

Also note that mark() misuses the error code 'R_INSTALL_ERROR_FAILED'
for reporting a status file update failure.
The code actually belongs to the install module and should indicate a
broken installation.

Thus, a better solution is to not let mark_active() fail for slot status
update errors but emit a simple log message instead.
Since we do not have to differentiate error codes then, it also makes
the calling codes much cleaner.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>