src/install: refactor boot slot marking during installation
An essential part of the installation process is to properly mark the
just-installed slot group as 'bad' for the bootloader before starting
the slot writing and to mark it 'active' again after having updated all
slots.

The slot used for marking 'bad' and 'active' must not change during an
installation and should be known before actually starting the
installation. Thus refactor the redundant determination into a single
method 'get_boot_mark_slot()' to be called right after having determined
the installation plans.

Note that since we can also ship bundles with images for slots that are
not part of any of the redundant slot groups, slot marking is not
mandatory and thus get_boot_mark_slot() can also return NULL.

The refactoring also fixes a potential long-lasting bug in slot marking
since the original code allowed to mark multiple slots. However, since
there is only one bootloader that can only boot one slot, this does not
make much sense and could break setups.
But, since no one ever reported having run into a situation it might be
that it is sufficiently prevented by other sanity checking before.
So we are lazy here and simply emit a g_error() (with a helpful error
message) instead of implementing GError handling.

Another minor issue this fixes is that the

| Leaving target slot non-bootable as requested by activate_installed == false.

message is no longer emitted for cases where no slot to mark is
available at all.

The actual marking of the valid slot is not changed by this refactoring.

Replaces #362.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>