src/config_file: emphasize importance of valid statusfile settings
If statusfile is unset, be more verbose and mention 'data-directory'
option which did not exist yet when the existing message was
introduced.

Could also be turned into a warning later on.

If per-slot slot status is used (either selected or as fallback), note
that system status information is not supported.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>