test/install: do not use g_idle_add() for r_quit()
Otherwise, when not having a main loop set up, a previous r_quit() call
will be queued up until we reach the thread tests that actually set up a
main loop.

There it leads to an early abort of the main loop and thus freeing
memory after usage will not work.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>