src/bootchooser: support the changed output of efibootmgr 18
Since efibootmgr 18, the default output of `efibootmgr` is now more
verbose [1], which breaks the assumptions made by RAUC in regards to
parsing the output. This issue is also affecting others [2].

Fix the parsing logic by unconditionally removing anything after a tab
in the detected name part of the boot entry, so that the unused data
part is discarded.
Emit a debug message for each found boot entry, as this helps in
detecting issues in the future.

[1] https://github.com/rhboot/efibootmgr/commit/8ec3e9dedb3cb62f19847794012420b90f475398
[2] https://github.com/rhboot/efibootmgr/issues/169

Signed-off-by: David Runge <dave@sleepmap.de>