README: clarify license
Simply including the LGPL license text is not enough information to
determine the license of a software project, as Section 13 of the LGPL
allows licensees to choose the terms of "any later version of the
license" if specified so by the copyright holders.

The information in meson.build implies "LGPL-2.1-or-later" as project
license. As a minimal measure to make this information explicit, add the
usual license notice and warranty disclaimer as recommended in the GNU
GPL How-To to the README file, and also add a minimal copyright notice
as required by Section 1 of the LGPL.

Let the LGPL badge at the start of the README point to the relevant
section further down.

Link: https://www.gnu.org/licenses/gpl-howto.html#why-license-notices
Link: https://www.gnu.org/licenses/identify-licenses-clearly.html
Signed-off-by: Roland Hieber <rhi@pengutronix.de>