src/manifest: ignore new per-image keys version, description and build
In some cases, it's useful to annotate images with more specific information
than we currently have at the bundle level. For now, just use the same keys as
at the bundle level.

In this change, we just ignore them to allow installation of bundles which
contain them in the manifest.

Later, we should store this information in the slot status and expose it via
D-Bus and 'rauc status'.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>