github: workflows: coverity: don't complain about missing secrets
On forked repositories, the coverity token is usually not configured in
the secrets, so running the coverity workflow will fail. Since it is
configured to run on a nightly cron schedule, this will send one useless
e-mail per night to the owner of the fork.

Run the workflow only if the access token is configured.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>