src/main: correctly use intern strings for slot status generation
In contrast to other string put into the RaucSlot struct, the 'name' and
'sclass' members are expected to be intern strings.
Thus they are not freed in r_slot_free().

When creating the slot struct in retrieve_slot_states_via_dbus() we
should have this in mind and also use only intern strings instead of
normal heap-allocated ones that would not be freed later.

To make this more obvious, also add a comment within the free function
itself.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>