src: detect config file change and warn in service log
A common usage issue with RAUC is to change the system configuration but
not to restart the rauc service to re-read the configuration.

To make this more obvious, at least in the service log output, this
implements a simple checksum-based checking for changes in the
system.conf.

On config file loading, the checksum of the original parsed
configuration is cached and used as reference for later comparisons of
re-calculated checksums.

The checker function r_config_file_modified_check() that implements the
comparison is used first of all in the service on calls of 'install' or
'status' actions and prints a warning.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>