src/main: fix --no-verify usage for rauc resign
In 3272c7ba ("src/main: support --no-verify for rauc resign"),
the --no-verify argument was added to 'rauc resign' to allow resigning a
bundle without the need to verify against the keyring.

Due to the additional sanity check in src/main, it was however not
possible to omit the keyring and thus this implicitly enforces a
(possibly unexpected) post-signing verification.

A check for a missing keyring is implemented in check_bundle() anyway,
thus the extra check in resign_start() is safe to omit.

To ensure this is also covered by testing, we simply omit the --keyring
argument in the no-verify resigning test.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>