src/bundle: remove 'manifest' arg from decrypt_bundle_payload()
We provided a bundle and a manifest as argument to
decrypt_bundle_payload(), but for crypt bundles, the manifest is part of
the bundle structure anyway.

We also made use of this at some points (i.e. used 'bundle->manifest') but
at other code lines we just used 'manifest'.
Thus get rid of this inconsistency and make it less prone to error by
just removing the 'manifest' argument.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>