src/bundle: do not re-create verity hash in resign_bundle()
The bundle resigning should not change the bundle content.

But so far, a resign_bundle() called the sign_bundle() method to
generate the signature. However, the sign_bundle() method not only
creates a signature but also creates and stores the verity hash
information in the bundle (for non-plain bundles).

As a result, the manifest and thus also the 'bundle hash' will be
different after a 'rauc resign' which is not the intended behavior.

Fix this by using the newly-introduced generate_bundle_signature()
method (together with append_signature_to_bundle()) to just generate the
signature without modifying the verity hash.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>