src/bundle: no need to check bundle payload in replace_signature()
Checking the bundle content is not required when not actually accessing
it. The only place to access the bundle content is when loading the
manifest from the bundle in case of a 'plain' format bundle.
But for this there already is a distinct call of check_bundle_payload()
in load_manifest_from_bundle().

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>