src/mark: ensure that selected slot is bootable
We should not call the bootchooser backend for non-bootable slots and
give the caller early feedback, thus only find slots if they have a
proper bootname property set and error-out for those that have not.

The check for parent instead is quite useless here, thus remove it.

Since we also do not need to re-ensure that a slot has a proper class,
just use g_hash_table_lookup() instead of the former lookup while loop.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>