src/bundle: remove signature size sanity check
Historically, checking that the signature size is smaller than the
bundle size was a sensible test.
However, with the added support for encrypted bundles, we may run in
cases where the bundle size is actually smaller than the signature size.
This happens because the CMS structure holds the individual recipients
for encrypted bundles and with having thousands of recipients, the CMS
structure size can reach a notable size. If that's combined with a quite
small update image, the sanity check can be triggered and prevents from
installing such a bundle.

Since this is just a sanity check, since we have other sanity checks and
since we intend to support large amounts of encryption recipients,
simply remove the sanity check.

Fixes #1308.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>