src: rename 'codesign' check purpose to 'codesign-rauc'
OpenSSL so far defined a number of certificate check purposes but none
for 'codesign'. Thus RAUC added its own purpose check for this.

Since v3.2.0, OpenSSL now defines a 'codesign' check purpose itself
which not only conflicts with the RAUC variant in naming, but also has
some differences compared to the RAUC check:

* OpenSSL enforces Key Usage existence, while RAUC only checks that it
  allows digital signatures if it exists
* RAUC enforces that Extended Key Usage on CA certs allows code signing
  if it exists

While RAUC is stricter on the one hand and OpenSSL on the other hand, we
tend to preserve the RAUC check first of all.
For this, the RAUC check purpose is renamed to 'codesign-rauc' and the
config file setting 'check-purpose=codesign' will be mapped to this.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>