.github/workflows/tests: use podman instead of docker and custom contaners
We need to use podman from Ubuntu noble to properly support multiarch manifests
with architecture variants (arm/v5, arm/v7) for our own ci containers. By using
the prebuilt containers, we avoid the slow installation of dependencies for
every job.

Also run the tests as a user inside the container, as this automatically skips
tests that cannot work in a container.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>