docs/using: be more explicit on current pre-install hook implementation
Pre-install hooks currently guarantee that the target slot is mounted.
For this, we explicitly error-out the update installtion if the target
slot is unmountable for any reason.

This can cause unexpected troubles when running this from a system that
does not have the target slot formatted (yet) or where the target slot's
file system is broken for any reason.

Be more explicit in the documentation about this fact. We'll likely
need a way to handle mount errors more gracefully in the future.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>