src/main: more explicit handling of rauc install user interruption
When running 'rauc install ..' and pressing Ctrl+C, a common expectation
is that this will interrupt the ongoing installation.

However, there is no current support for this in RAUC and it is also not
really trivial.

Thus, to at least prevent wrong expectations, handle SIGINT explicitly
and emit a message to the user to inform about the ongoing installation.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>