src/nbd: manually trigger NBD disconnect
When invoking 'rauc info' the connection to the NBD server is never
handed to the kernel and thus just terminated by the code in
r_nbd_stop_server() that just closes the socket.

This leads to messages like

| rauc-Message: 23:00:27.704: nbd server failed with: failed to read request from client: Unexpected end of file
| rauc-nbd-Message: 23:00:27.714: failed to stop ndb server: failed to stop streaming subprocess: Child process exited with code 1

This can be handled better by manually invoking a nbd disconnect before
that triggers a shutdown on NBD server side that prevents the error
message and also prints stats etc.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>