test/rauc.t: test faketime before relying on it
On some platforms faketime doesn't work as intended due to not
supporting the libc functions that provide 64bit time_t on 32bit
architectures.

This problem already made rauc fail to build in Debian unstable, see
https://bugs.debian.org/1032177. With the expectation that the problem
will also affect rauc mainline at some time in the future,
precautionarily check faketime being functional before using it.

Note that the date string is parsed and emitted in the local timezone,
so it cannot happen that with a working faketime a date from 2017 is
printed.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>