src/install: fix slot skipping detection for global slot status
In 67ccd898 ("src/install: ensure invalidating slot status before
updating"), we clear the slot checksum in the status file before
updating the slot. In case of an installation abort, this should avoid
having a valid-looking checksum in the status file.

However, this handling clears the 'slot_state->checksum.digest' right
before we use this in the slot skipping check expression.
As a result, slot skipping is broken and the image will always be
written, regardless of the 'install-same' setting.

Fix this by just moving the checksum clearing below the slot skipping
check.

Fixes #1158, first released with v1.11.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>