ignore SIGPIPE to handle broken pipes explicitly
Glib already sets SIGPIPE to ignore when using gsocket and will do the
same for gsubprocess from 2.82 as well [1]. As we want to handle all
broken pipe errors explicitly, we should do it globally.

We add the call in r_context_conf instead of in main() to catch the
relevant tests as well. For the nbd helper (which doesn't use a
context), we do it on startup.

[1] https://gitlab.gnome.org/GNOME/glib/-/merge_requests/3991

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>