src/utils: relax semantic version parsing
Change the string parsing to also allow "relaxed semantic version"
parsing where not the full Major.Minor.Patch is required, but also
variants with just Minor.Patch[-pre_release][+build] or
Patch[-pre_release][+build] are allowed.

Note that internally the core-version is still kept as as guint[3] but
the missing version parts are just filled in with zero.

Signed-off-by: Johannes Schneider <johannes.schneider@leica-geosystems.com>