Add per-slot option for casync seed device
As mentioned in #1200, using casync in blob mode with UBI volumes is
kind of pointless at the moment, because casync cannot be effective
without a seed.

The solution proposed in the comment in casync_extract_image sounds
good and would make this use case effortless from a user perspective,
but it would require some development effort to implement said logic.

This solution instead leaves the mapping logic to the user, which is
likely pretty simple or even nothing at all (e.g. in the case of a
squashfs root filesystem, which by necessity already has a ubiblock).

A typical configuration would look like this:

[slot.rootfs.0]
device=/dev/ubi0_0
casync-seed-device=/dev/ubiblock0_0
type=ubivol

Of course it is still possible for someone to implement the ubiblock
mapping logic in RAUC in the future if they see a need for it. But with
this change at least the use case is viable with minimal integration
work.

Signed-off-by: Tim van der Staaij <git@tim.vanderstaaij.email>